Tag real-DB and embedded-asset tests as integration tests
