+++ internal/game/game_test.go
@@ -9,12 +9,9 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
-	_ "modernc.org/sqlite"
 
-	"github.com/starquake/topbanana/internal/dbtest"
 	. "github.com/starquake/topbanana/internal/game"
 	"github.com/starquake/topbanana/internal/quiz"
-	"github.com/starquake/topbanana/internal/store"
 )
 
 var errStub = errors.New("stub error")
@@ -246,74 +243,6 @@ func (stubQuizStore) MoveQuestionToRound(_ context.Context, _, _, _ int64) error
 	return errStub
 }
 
-// seededAdminID is the id of the admin row inserted by migration
-// 20260111110308_add_admin_player.sql. Quiz fixtures attribute
-// themselves to this admin so the NOT NULL created_by_player_id
-// column from migration 20260520200000 (#281) is satisfied.
-const seededAdminID int64 = 1
-
-func newTestQuiz(t *testing.T) *quiz.Quiz {
-	t.Helper()
-
-	return &quiz.Quiz{
-		Title:             "Flurpsydurpsy",
-		Slug:              "flurpsydurpsy",
-		CreatedByPlayerID: seededAdminID,
-		Questions: []*quiz.Question{
-			{
-				Text:     "What is the capital of France?",
-				Position: 10,
-				Options: []*quiz.Option{
-					{Text: "Paris", Correct: true},
-					{Text: "London"},
-				},
-			},
-			{
-				Text:     "What is the capital of Germany?",
-				Position: 20,
-				Options: []*quiz.Option{
-					{Text: "Berlin", Correct: true},
-					{Text: "Hamburg"},
-				},
-			},
-			{
-				Text:     "What is the capital of Spain?",
-				Position: 30,
-				Options: []*quiz.Option{
-					{Text: "Madrid", Correct: true},
-					{Text: "Barcelona"},
-				},
-			},
-		},
-	}
-}
-
-// assertBoundaryWindow pins the #548 auto-advance contract at the
-// service layer: a round-boundary item carries a non-zero
-// StartedAt/ExpiredAt window exactly one quiz-default answer duration
-// (timeLimitSeconds) long, for both phases.
-func assertBoundaryWindow(t *testing.T, item *Item, timeLimitSeconds int) {
-	t.Helper()
-	if item.StartedAt.IsZero() {
-		t.Error("item.StartedAt is zero, want a populated timestamp")
-	}
-	if item.ExpiredAt.IsZero() {
-		t.Error("item.ExpiredAt is zero, want a populated timestamp")
-	}
-	want := time.Duration(timeLimitSeconds) * time.Second
-	if got := item.ExpiredAt.Sub(item.StartedAt); got != want {
-		t.Errorf("item window ExpiredAt-StartedAt = %v, want %v (quiz default)", got, want)
-	}
-}
-
-func newTestGame(t *testing.T, qz *quiz.Quiz) *Game {
-	t.Helper()
-
-	return &Game{
-		QuizID: qz.ID,
-	}
-}
-
 func TestGame_IsCompleted(t *testing.T) {
 	t.Parallel()
 
@@ -452,765 +381,6 @@ func TestGame_HasOpenQuestion(t *testing.T) {
 		}
 		if got, want := g.HasOpenQuestion(), true; got != want {
 			t.Errorf("HasOpenQuestion() = %v, want %v (last question must be resumable)", got, want)
-		}
-	})
-}
-
-func TestService_GetGameForPlayerOnQuiz(t *testing.T) {
-	t.Parallel()
-
-	t.Run("returns existing game with quiz populated", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		const playerID = int64(1)
-		created, err := svc.CreateGame(ctx, testQuiz.ID, playerID)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		resumed, err := svc.GetGameForPlayerOnQuiz(ctx, playerID, testQuiz.ID)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if got, want := resumed.ID, created.ID; got != want {
-			t.Errorf("resumed.ID = %q, want %q", got, want)
-		}
-		if resumed.Quiz == nil {
-			t.Fatal("resumed.Quiz is nil, want populated quiz")
-		}
-		if got, want := resumed.Quiz.ID, testQuiz.ID; got != want {
-			t.Errorf("resumed.Quiz.ID = %d, want %d", got, want)
-		}
-		// IsCompleted should work because Quiz is populated.
-		if got, want := resumed.IsCompleted(), false; got != want {
-			t.Errorf("IsCompleted() = %v, want %v (no questions issued yet)", got, want)
-		}
-	})
-
-	t.Run("returns ErrGameNotFound when player has no game", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		_, err := svc.GetGameForPlayerOnQuiz(ctx, 999, testQuiz.ID)
-		if got, want := err, ErrGameNotFound; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("returns ErrQuizNotFound when quiz missing", func(t *testing.T) {
-		t.Parallel()
-
-		svc := NewService(stubStore{}, stubQuizStore{
-			getQuiz: func(_ context.Context, _ int64) (*quiz.Quiz, error) {
-				return nil, quiz.ErrQuizNotFound
-			},
-		}, slog.New(slog.DiscardHandler))
-
-		_, err := svc.GetGameForPlayerOnQuiz(t.Context(), 1, 999)
-		if got, want := err, quiz.ErrQuizNotFound; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-}
-
-func TestService_CreateGame_RejectsSecondAttempt(t *testing.T) {
-	t.Parallel()
-
-	ctx := t.Context()
-	db := dbtest.Open(t)
-
-	quizStore := store.NewQuizStore(db, slog.Default())
-	gameStore := store.NewGameStore(db, slog.Default())
-
-	testQuiz := newTestQuiz(t)
-	if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-		t.Fatalf("failed to create quiz: %v", err)
-	}
-
-	svc := NewService(gameStore, quizStore, slog.Default())
-
-	const playerID = int64(1)
-	if _, err := svc.CreateGame(ctx, testQuiz.ID, playerID); err != nil {
-		t.Fatalf("failed to create initial game: %v", err)
-	}
-
-	// Second attempt for the same (player, quiz) must be rejected.
-	_, err := svc.CreateGame(ctx, testQuiz.ID, playerID)
-	if got, want := err, ErrGameAlreadyExists; !errors.Is(got, want) {
-		t.Errorf("err = %v, want %v", got, want)
-	}
-}
-
-func TestService_ResetGamesForPlayerOnQuiz(t *testing.T) {
-	t.Parallel()
-
-	t.Run("reset clears existing game and lets a fresh CreateGame succeed", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		const playerID = int64(1)
-		if _, err := svc.CreateGame(ctx, testQuiz.ID, playerID); err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		// Sanity check: GetGameForPlayerOnQuiz finds the game first.
-		if _, err := svc.GetGameForPlayerOnQuiz(ctx, playerID, testQuiz.ID); err != nil {
-			t.Fatalf("expected game to exist before reset: %v", err)
-		}
-
-		if err := svc.ResetGamesForPlayerOnQuiz(ctx, playerID, testQuiz.ID); err != nil {
-			t.Fatalf("ResetGamesForPlayerOnQuiz err = %v, want nil", err)
-		}
-
-		_, err := svc.GetGameForPlayerOnQuiz(ctx, playerID, testQuiz.ID)
-		if got, want := err, ErrGameNotFound; !errors.Is(got, want) {
-			t.Errorf("after reset, err = %v, want %v", got, want)
-		}
-
-		if _, err = svc.CreateGame(ctx, testQuiz.ID, playerID); err != nil {
-			t.Errorf("CreateGame after reset err = %v, want nil", err)
-		}
-	})
-
-	t.Run("idempotent - calling reset twice is fine", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		const playerID = int64(1)
-		if _, err := svc.CreateGame(ctx, testQuiz.ID, playerID); err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		if err := svc.ResetGamesForPlayerOnQuiz(ctx, playerID, testQuiz.ID); err != nil {
-			t.Fatalf("first reset err = %v, want nil", err)
-		}
-		if err := svc.ResetGamesForPlayerOnQuiz(ctx, playerID, testQuiz.ID); err != nil {
-			t.Errorf("second reset err = %v, want nil (reset must be idempotent)", err)
-		}
-	})
-
-	t.Run("returns ErrQuizNotFound when quiz missing", func(t *testing.T) {
-		t.Parallel()
-
-		svc := NewService(stubStore{}, stubQuizStore{
-			quizExists: func(_ context.Context, _ int64) (bool, error) {
-				return false, nil
-			},
-		}, slog.New(slog.DiscardHandler))
-
-		err := svc.ResetGamesForPlayerOnQuiz(t.Context(), 1, 999)
-		if got, want := err, quiz.ErrQuizNotFound; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-}
-
-func TestService_SubmitAnswer(t *testing.T) {
-	t.Parallel()
-
-	t.Run("rejects option from a different question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		g, err := svc.CreateGame(ctx, testQuiz.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		gq, err := svc.GetNextQuestion(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		// Use a correct option from the second question (different from the active question).
-		wrongQuestionOption := testQuiz.Questions[1].Options[0] // Berlin, Correct: true, but for question 2
-
-		_, err = svc.SubmitAnswer(ctx, g.ID, 1, gq.QuizQuestion.ID, wrongQuestionOption.ID, time.Time{})
-		if got, want := err, ErrOptionNotInQuestion; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("accepts option belonging to the active question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		g, err := svc.CreateGame(ctx, testQuiz.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		gq, err := svc.GetNextQuestion(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		correctOption := testQuiz.Questions[0].Options[0] // Paris, Correct: true
-
-		_, err = svc.SubmitAnswer(ctx, g.ID, 1, gq.QuizQuestion.ID, correctOption.ID, time.Time{})
-		if err != nil {
-			t.Errorf("unexpected error: %v", err)
-		}
-	})
-}
-
-func TestService_GetResults(t *testing.T) {
-	t.Parallel()
-
-	t.Run("returns player with highest score as winner", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		g, err := svc.CreateGame(ctx, testQuiz.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		gq, err := svc.GetNextQuestion(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		insertPlayer2 := `INSERT INTO players (id, display_name, email, created_at) VALUES (2, 'player2', 'player2@test.com', CURRENT_TIMESTAMP)`
-		if _, err = db.ExecContext(ctx, insertPlayer2); err != nil {
-			t.Fatalf("failed to insert player 2: %v", err)
-		}
-		// Participant gate (#272): player 2 needs an explicit
-		// participant row, otherwise SubmitAnswer rejects them as a
-		// non-participant. The bug-fix for #272 made the gate strict;
-		// pre-fix this test inadvertently relied on the missing check.
-		if err = gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: g.ID, PlayerID: 2, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("failed to create participant for player 2: %v", err)
-		}
-
-		correctOption := testQuiz.Questions[0].Options[0] // Paris, Correct: true
-		wrongOption := testQuiz.Questions[0].Options[1]   // London, Correct: false
-
-		if _, err = svc.SubmitAnswer(ctx, g.ID, 1, gq.QuizQuestion.ID, correctOption.ID, time.Time{}); err != nil {
-			t.Fatalf("failed to submit answer for player 1: %v", err)
-		}
-		if _, err = svc.SubmitAnswer(ctx, g.ID, 2, gq.QuizQuestion.ID, wrongOption.ID, time.Time{}); err != nil {
-			t.Fatalf("failed to submit answer for player 2: %v", err)
-		}
-
-		results, err := svc.GetResults(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get results: %v", err)
-		}
-
-		if got, want := results.Winner, int64(1); got != want {
-			t.Errorf("Winner = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("returns no winner on tie", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		g, err := svc.CreateGame(ctx, testQuiz.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-
-		gq, err := svc.GetNextQuestion(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		insertPlayer2 := `INSERT INTO players (id, display_name, email, created_at) VALUES (2, 'player2', 'player2@test.com', CURRENT_TIMESTAMP)`
-		if _, err = db.ExecContext(ctx, insertPlayer2); err != nil {
-			t.Fatalf("failed to insert player 2: %v", err)
-		}
-		// Participant gate (#272): player 2 needs an explicit
-		// participant row, otherwise SubmitAnswer rejects them as a
-		// non-participant. The bug-fix for #272 made the gate strict;
-		// pre-fix this test inadvertently relied on the missing check.
-		if err = gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: g.ID, PlayerID: 2, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("failed to create participant for player 2: %v", err)
-		}
-
-		wrongOption := testQuiz.Questions[0].Options[1] // London, Correct: false
-
-		if _, err = svc.SubmitAnswer(ctx, g.ID, 1, gq.QuizQuestion.ID, wrongOption.ID, time.Time{}); err != nil {
-			t.Fatalf("failed to submit answer for player 1: %v", err)
-		}
-		if _, err = svc.SubmitAnswer(ctx, g.ID, 2, gq.QuizQuestion.ID, wrongOption.ID, time.Time{}); err != nil {
-			t.Fatalf("failed to submit answer for player 2: %v", err)
-		}
-
-		results, err := svc.GetResults(ctx, g.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get results: %v", err)
-		}
-
-		if got, want := results.Winner, int64(0); got != want {
-			t.Errorf("Winner = %v, want %v (expected no winner on tie)", got, want)
-		}
-	})
-
-	t.Run("skips answers whose option was deleted without panicking", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-
-		// A finished game with one answer that references option 99,
-		// which GetOptionsByIDs no longer returns (the option row was
-		// deleted out from under the answer). The score loop must skip
-		// it rather than dereference a nil Option.
-		gameWithDanglingAnswer := &Game{
-			ID:           "game-deleted-opt",
-			Participants: []*Participant{{PlayerID: 1}},
-			Questions: []*Question{
-				{
-					ID: 1,
-					Answers: []*Answer{
-						{PlayerID: 1, OptionID: 99, AnsweredAt: time.Now()},
-					},
-				},
-			},
-		}
-
-		gs := stubStore{
-			getGame: func(_ context.Context, _ string) (*Game, error) {
-				return gameWithDanglingAnswer, nil
-			},
-		}
-		qs := stubQuizStore{
-			getOptionsByIDs: func(_ context.Context, _ []int64) ([]*quiz.Option, error) {
-				return nil, nil
-			},
-		}
-
-		svc := NewService(gs, qs, slog.Default())
-
-		results, err := svc.GetResults(ctx, "game-deleted-opt", 1)
-		if err != nil {
-			t.Fatalf("GetResults err = %v, want nil", err)
-		}
-		if got, want := results.PlayerScores[1], 0; got != want {
-			t.Errorf("PlayerScores[1] = %d, want %d", got, want)
-		}
-	})
-}
-
-func TestService_GetNextQuestion(t *testing.T) {
-	t.Parallel()
-
-	t.Run("first question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-
-		var err error
-		err = quizStore.CreateQuiz(ctx, testQuiz)
-		if err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		err = gameStore.CreateGame(ctx, testGame)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-		// Participant gate (#272): the service rejects callers that
-		// aren't on the participant list. These tests bypass the
-		// service's CreateGame, so the participant row has to be
-		// seeded explicitly here.
-		if err = gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("failed to create participant: %v", err)
-		}
-
-		service := NewService(gameStore, quizStore, slog.Default())
-		gq, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		if gq == nil {
-			t.Fatal("expected gq to be non-nil")
-
-			return
-		}
-
-		if cmp.Diff(gq.QuizQuestion, testQuiz.Questions[0]) != "" {
-			t.Errorf("got qs: %+v, want %+v", gq.QuizQuestion, testQuiz.Questions[0])
-		}
-	})
-
-	t.Run("second question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-
-		var err error
-		err = quizStore.CreateQuiz(ctx, testQuiz)
-		if err != nil {
-			t.Fatalf("failed to create quiz: %v", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		err = gameStore.CreateGame(ctx, testGame)
-		if err != nil {
-			t.Fatalf("failed to create game: %v", err)
-		}
-		// Participant gate (#272): the service rejects callers that
-		// aren't on the participant list. These tests bypass the
-		// service's CreateGame, so the participant row has to be
-		// seeded explicitly here.
-		if err = gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("failed to create participant: %v", err)
-		}
-
-		err = gameStore.CreateQuestion(ctx, &Question{GameID: testGame.ID, QuestionID: testQuiz.Questions[0].ID})
-		if err != nil {
-			t.Fatalf("failed to create game question: %v", err)
-		}
-
-		service := NewService(gameStore, quizStore, slog.Default())
-		gq, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("failed to get next question: %v", err)
-		}
-
-		if gq == nil {
-			t.Fatal("expected gq to be non-nil")
-
-			return
-		}
-
-		if cmp.Diff(gq.QuizQuestion, testQuiz.Questions[1]) != "" {
-			t.Errorf("got qs: %+v, want %+v", gq.QuizQuestion, testQuiz.Questions[1])
-		}
-	})
-
-	t.Run("started_at sits in the future to honour the reveal delay", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		// Participant gate (#272): seed the participant directly since
-		// these tests bypass Service.CreateGame.
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		service := NewService(gameStore, quizStore, slog.Default())
-		issuedAt := time.Now()
-		gq, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNextQuestion err = %v, want nil", err)
-		}
-
-		// StartedAt must be at least 2 seconds in the future relative
-		// to issuedAt - the 3s reveal delay (#247) gives the player
-		// time to read the question before the answer window opens.
-		// 2s lower bound is forgiving of clock granularity on the
-		// test machine; the production constant is 3s.
-		if got, lower := gq.StartedAt.Sub(issuedAt), 2*time.Second; got < lower {
-			t.Errorf("StartedAt - issuedAt = %v, want >= %v (reveal delay)", got, lower)
-		}
-		// ExpiredAt sits one answer window further. The window is 10s
-		// (the unexported defaultExpiration constant); duplicated here
-		// as a literal so this external_test file doesn't have to
-		// reach into package internals.
-		if got, want := gq.ExpiredAt.Sub(gq.StartedAt), 10*time.Second; got != want {
-			t.Errorf("ExpiredAt - StartedAt = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("back-to-back calls return the same in-flight question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		// Participant gate (#272): seed the participant directly since
-		// these tests bypass Service.CreateGame.
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		service := NewService(gameStore, quizStore, slog.Default())
-		first, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("first GetNextQuestion err = %v, want nil", err)
-		}
-
-		// Second call without submitting an answer must return the same
-		// game_questions row - same ID, same timing anchors - so a
-		// mid-question reload doesn't skip the question.
-		second, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("second GetNextQuestion err = %v, want nil", err)
-		}
-		if got, want := second.ID, first.ID; got != want {
-			t.Errorf("second.ID = %d, want %d (resume must hand back same row)", got, want)
-		}
-		if got, want := second.QuizQuestion.ID, first.QuizQuestion.ID; got != want {
-			t.Errorf("second.QuizQuestion.ID = %d, want %d", got, want)
-		}
-		if got, want := second.StartedAt, first.StartedAt; !got.Equal(want) {
-			t.Errorf("second.StartedAt = %v, want %v (timing anchor must not reset)", got, want)
-		}
-		if got, want := second.ExpiredAt, first.ExpiredAt; !got.Equal(want) {
-			t.Errorf("second.ExpiredAt = %v, want %v", got, want)
-		}
-
-		// And no extra game_questions row was inserted.
-		g, err := gameStore.GetGame(ctx, testGame.ID)
-		if err != nil {
-			t.Fatalf("GetGame err = %v, want nil", err)
-		}
-		if got, want := len(g.Questions), 1; got != want {
-			t.Errorf("game_questions count = %d, want %d (resume must not insert)", got, want)
-		}
-	})
-
-	t.Run("advance after an expired unanswered question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		// Seed an unanswered game_question whose answer window has
-		// already closed - the timeout path leaves rows like this.
-		// The advance branch must move past it instead of pinning the
-		// player on the expired question.
-		past := time.Now().Add(-1 * time.Minute)
-		expired := &Question{
-			GameID:     testGame.ID,
-			QuestionID: testQuiz.Questions[0].ID,
-			StartedAt:  past,
-			ExpiredAt:  past.Add(10 * time.Second),
-		}
-		if err := gameStore.CreateQuestion(ctx, expired); err != nil {
-			t.Fatalf("CreateQuestion err = %v, want nil", err)
-		}
-
-		service := NewService(gameStore, quizStore, slog.Default())
-		gq, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNextQuestion err = %v, want nil", err)
-		}
-		if got, want := gq.QuizQuestion.ID, testQuiz.Questions[1].ID; got != want {
-			t.Errorf("advanced to QuizQuestion.ID = %d, want %d (expired Q1 must not pin)", got, want)
-		}
-	})
-
-	t.Run("SetRevealDelay shrinks the reveal-to-answer gap", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		// Participant gate (#272): seed the participant directly since
-		// these tests bypass Service.CreateGame.
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		// Sub-second reveal mirrors the e2e config: shorter than the
-		// default 3s but still leaves the reveal phase observable.
-		service := NewService(gameStore, quizStore, slog.Default())
-		service.SetRevealDelay(200 * time.Millisecond)
-		issuedAt := time.Now()
-		gq, err := service.GetNextQuestion(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNextQuestion err = %v, want nil", err)
-		}
-
-		// StartedAt should sit close to (issuedAt + 200ms). Generous
-		// upper bound to absorb scheduler jitter on busy CI runners.
-		if got, upper := gq.StartedAt.Sub(issuedAt), 1*time.Second; got >= upper {
-			t.Errorf("StartedAt - issuedAt = %v, want < %v (override should shrink reveal)", got, upper)
 		}
 	})
 }
@@ -1824,383 +994,6 @@ func TestService_GetQuizLeaderboard(t *testing.T) {
 	})
 }
 
-// firstRoundID returns the id of the only round a freshly created quiz
-// has - the default "Round 1" the store stamps on CreateQuiz (#444).
-func firstRoundID(t *testing.T, qz *quiz.Quiz, quizStore *store.QuizStore) int64 {
-	t.Helper()
-	rounds, err := quizStore.ListRoundsByQuiz(t.Context(), qz.ID)
-	if err != nil {
-		t.Fatalf("ListRoundsByQuiz err = %v, want nil", err)
-	}
-	if len(rounds) == 0 {
-		t.Fatal("quiz has no rounds, want at least the default")
-	}
-
-	return rounds[0].ID
-}
-
-// giveRoundSummary stamps a summary on an existing round so its boundary
-// fires during play. The play iterator skips a round whose summary is
-// empty (#444), so boundary-emission tests must author one first.
-func giveRoundSummary(t *testing.T, quizStore *store.QuizStore, roundID int64, summary string) {
-	t.Helper()
-	round, err := quizStore.GetRound(t.Context(), roundID)
-	if err != nil {
-		t.Fatalf("GetRound err = %v, want nil", err)
-	}
-	round.Summary = summary
-	if err := quizStore.UpdateRound(t.Context(), round); err != nil {
-		t.Fatalf("UpdateRound err = %v, want nil", err)
-	}
-}
-
-func TestService_GetNext(t *testing.T) {
-	t.Parallel()
-
-	t.Run("returns the first question of the first round", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		item, err := svc.GetNext(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNext err = %v, want nil", err)
-		}
-		if got, want := item.Type, ItemTypeQuestion; got != want {
-			t.Errorf("item.Type = %q, want %q", got, want)
-		}
-		if got, want := item.Question.QuizQuestion.ID, testQuiz.Questions[0].ID; got != want {
-			t.Errorf("item.Question.QuizQuestion.ID = %d, want %d", got, want)
-		}
-	})
-
-	t.Run("emits the results boundary after every question in the round is issued", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		roundID := firstRoundID(t, testQuiz, quizStore)
-		giveRoundSummary(t, quizStore, roundID, "Round one wrapped up")
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		// Issue every question and mark the intro phase seen so the only
-		// remaining item is the round's results boundary.
-		for _, q := range testQuiz.Questions {
-			if err := gameStore.CreateQuestion(
-				ctx, &Question{GameID: testGame.ID, QuestionID: q.ID},
-			); err != nil {
-				t.Fatalf("CreateQuestion err = %v, want nil", err)
-			}
-		}
-		if err := gameStore.MarkRoundSeen(ctx, testGame.ID, roundID, RoundPhaseIntro); err != nil {
-			t.Fatalf("MarkRoundSeen (intro) err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		item, err := svc.GetNext(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNext err = %v, want nil", err)
-		}
-		if got, want := item.Type, ItemTypeRoundBoundary; got != want {
-			t.Errorf("item.Type = %q, want %q", got, want)
-		}
-		if got, want := item.Phase, RoundPhaseResults; got != want {
-			t.Errorf("item.Phase = %q, want %q", got, want)
-		}
-		if got, want := item.Round.ID, roundID; got != want {
-			t.Errorf("item.Round.ID = %d, want %d", got, want)
-		}
-		if got, want := item.Total, len(testQuiz.Questions); got != want {
-			t.Errorf("item.Total = %d, want %d", got, want)
-		}
-		if got, want := item.RoundQuestions, len(testQuiz.Questions); got != want {
-			t.Errorf("item.RoundQuestions = %d, want %d", got, want)
-		}
-		assertBoundaryWindow(t, item, testQuiz.TimeLimitSeconds)
-	})
-
-	t.Run("emits the intro boundary before the round's first question", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		roundID := firstRoundID(t, testQuiz, quizStore)
-		giveRoundSummary(t, quizStore, roundID, "Round one ahead")
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		item, err := svc.GetNext(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("GetNext err = %v, want nil", err)
-		}
-		if got, want := item.Type, ItemTypeRoundBoundary; got != want {
-			t.Errorf("item.Type = %q, want %q", got, want)
-		}
-		if got, want := item.Phase, RoundPhaseIntro; got != want {
-			t.Errorf("item.Phase = %q, want %q", got, want)
-		}
-		if got, want := item.Round.Summary, "Round one ahead"; got != want {
-			t.Errorf("item.Round.Summary = %q, want %q", got, want)
-		}
-		assertBoundaryWindow(t, item, testQuiz.TimeLimitSeconds)
-	})
-
-	t.Run("skips a round boundary the player has already seen", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		roundID := firstRoundID(t, testQuiz, quizStore)
-		giveRoundSummary(t, quizStore, roundID, "Round one wrapped up")
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-		for _, q := range testQuiz.Questions {
-			if err := gameStore.CreateQuestion(
-				ctx, &Question{GameID: testGame.ID, QuestionID: q.ID},
-			); err != nil {
-				t.Fatalf("CreateQuestion err = %v, want nil", err)
-			}
-		}
-		if err := gameStore.MarkRoundSeen(ctx, testGame.ID, roundID, RoundPhaseIntro); err != nil {
-			t.Fatalf("MarkRoundSeen (intro) err = %v, want nil", err)
-		}
-		if err := gameStore.MarkRoundSeen(ctx, testGame.ID, roundID, RoundPhaseResults); err != nil {
-			t.Fatalf("MarkRoundSeen (results) err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		_, err := svc.GetNext(ctx, testGame.ID, 1)
-		if got, want := err, ErrNoMoreQuestions; !errors.Is(got, want) {
-			t.Errorf("GetNext err = %v, want %v (all questions issued, both phases seen)", got, want)
-		}
-	})
-
-	t.Run("walks rounds in position order issuing each round's questions then its boundary", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		// Two rounds: round 1 holds Q1, round 2 holds Q2, each with a
-		// summary. The walk should be intro(r1), Q1, results(r1),
-		// intro(r2), Q2, results(r2). Each question is answered before
-		// advancing so the resume path does not hand the same in-flight
-		// question back; each boundary phase is acked so it is not
-		// re-emitted.
-		testQuiz := &quiz.Quiz{
-			Title:             "Two rounds",
-			Slug:              "two-rounds",
-			CreatedByPlayerID: seededAdminID,
-			Questions: []*quiz.Question{
-				{Text: "Q1", Position: 10, Options: []*quiz.Option{{Text: "a", Correct: true}, {Text: "b"}}},
-				{Text: "Q2", Position: 20, Options: []*quiz.Option{{Text: "c", Correct: true}, {Text: "d"}}},
-			},
-		}
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		round1 := firstRoundID(t, testQuiz, quizStore)
-		giveRoundSummary(t, quizStore, round1, "Round one wrapped up")
-		round2 := &quiz.Round{QuizID: testQuiz.ID, Position: 1, Title: "Round 2", Summary: "Round two wrapped up"}
-		if err := quizStore.CreateRound(ctx, round2); err != nil {
-			t.Fatalf("CreateRound err = %v, want nil", err)
-		}
-		if err := quizStore.MoveQuestionToRound(ctx, testQuiz.ID, testQuiz.Questions[1].ID, round2.ID); err != nil {
-			t.Fatalf("MoveQuestionToRound err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		answer := func(item *Item) {
-			t.Helper()
-			optID := item.Question.QuizQuestion.Options[0].ID
-			if _, err := svc.SubmitAnswer(
-				ctx, testGame.ID, 1, item.Question.QuizQuestion.ID, optID, time.Now(),
-			); err != nil {
-				t.Fatalf("SubmitAnswer err = %v, want nil", err)
-			}
-		}
-
-		nextBoundary := func(label string, roundID int64, phase RoundPhase) {
-			t.Helper()
-			item, err := svc.GetNext(ctx, testGame.ID, 1)
-			if err != nil {
-				t.Fatalf("GetNext (%s) err = %v, want nil", label, err)
-			}
-			if got, want := item.Type, ItemTypeRoundBoundary; got != want {
-				t.Fatalf("%s item.Type = %q, want %q", label, got, want)
-			}
-			if got, want := item.Phase, phase; got != want {
-				t.Fatalf("%s item.Phase = %q, want %q", label, got, want)
-			}
-			if got, want := item.Round.ID, roundID; got != want {
-				t.Errorf("%s boundary round = %d, want %d", label, got, want)
-			}
-			if err = svc.MarkRoundSeen(ctx, testGame.ID, 1, roundID, phase); err != nil {
-				t.Fatalf("MarkRoundSeen (%s) err = %v, want nil", label, err)
-			}
-		}
-		nextQuestion := func(label string, wantQuestionID int64) {
-			t.Helper()
-			item, err := svc.GetNext(ctx, testGame.ID, 1)
-			if err != nil {
-				t.Fatalf("GetNext (%s) err = %v, want nil", label, err)
-			}
-			if got, want := item.Type, ItemTypeQuestion; got != want {
-				t.Fatalf("%s item.Type = %q, want %q", label, got, want)
-			}
-			if got, want := item.Question.QuizQuestion.ID, wantQuestionID; got != want {
-				t.Errorf("%s question = %d, want %d", label, got, want)
-			}
-			answer(item)
-		}
-
-		nextBoundary("intro 1", round1, RoundPhaseIntro)
-		nextQuestion("Q1", testQuiz.Questions[0].ID)
-		nextBoundary("results 1", round1, RoundPhaseResults)
-		nextBoundary("intro 2", round2.ID, RoundPhaseIntro)
-		nextQuestion("Q2", testQuiz.Questions[1].ID)
-		nextBoundary("results 2", round2.ID, RoundPhaseResults)
-
-		// exhausted
-		_, err := svc.GetNext(ctx, testGame.ID, 1)
-		if got, want := err, ErrNoMoreQuestions; !errors.Is(got, want) {
-			t.Errorf("final GetNext err = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("resume returns the in-flight question unchanged", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-
-		first, err := svc.GetNext(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("first GetNext err = %v, want nil", err)
-		}
-		second, err := svc.GetNext(ctx, testGame.ID, 1)
-		if err != nil {
-			t.Fatalf("second GetNext err = %v, want nil", err)
-		}
-		if got, want := second.Type, ItemTypeQuestion; got != want {
-			t.Fatalf("second.Type = %q, want %q", got, want)
-		}
-		if got, want := second.Question.QuestionID, first.Question.QuestionID; got != want {
-			t.Errorf("resume question = %d, want %d (must hand back the same in-flight question)", got, want)
-		}
-	})
-}
-
 func TestNextRoundSlot(t *testing.T) {
 	t.Parallel()
 
@@ -2295,126 +1088,6 @@ func TestNextRoundSlot(t *testing.T) {
 		after := ExportNextRoundSlot(soloRounds, soloQuestions, map[int64]bool{3: true}, nil)
 		if got, want := after.Kind, ""; got != want {
 			t.Errorf("after Kind = %q, want %q (no boundary)", got, want)
-		}
-	})
-}
-
-func TestService_MarkRoundSeen(t *testing.T) {
-	t.Parallel()
-
-	t.Run("idempotent", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		roundID := firstRoundID(t, testQuiz, quizStore)
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		if err := svc.MarkRoundSeen(ctx, testGame.ID, 1, roundID, RoundPhaseIntro); err != nil {
-			t.Errorf("first MarkRoundSeen err = %v, want nil", err)
-		}
-		if err := svc.MarkRoundSeen(ctx, testGame.ID, 1, roundID, RoundPhaseIntro); err != nil {
-			t.Errorf("second MarkRoundSeen err = %v, want nil (idempotent)", err)
-		}
-	})
-
-	t.Run("unknown phase returns ErrInvalidRoundPhase", func(t *testing.T) {
-		t.Parallel()
-
-		svc := NewService(stubStore{}, stubQuizStore{}, slog.Default())
-		err := svc.MarkRoundSeen(t.Context(), "game-1", 1, 1, RoundPhase("bogus"))
-		if got, want := err, ErrInvalidRoundPhase; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("non-participant returns ErrGameNotFound", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-		roundID := firstRoundID(t, testQuiz, quizStore)
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		err := svc.MarkRoundSeen(ctx, testGame.ID, 999, roundID, RoundPhaseIntro)
-		if got, want := err, ErrGameNotFound; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
-		}
-	})
-
-	t.Run("round from a different quiz returns ErrRoundNotFound", func(t *testing.T) {
-		t.Parallel()
-
-		ctx := t.Context()
-		db := dbtest.Open(t)
-
-		quizStore := store.NewQuizStore(db, slog.Default())
-		gameStore := store.NewGameStore(db, slog.Default())
-
-		testQuiz := newTestQuiz(t)
-		if err := quizStore.CreateQuiz(ctx, testQuiz); err != nil {
-			t.Fatalf("CreateQuiz err = %v, want nil", err)
-		}
-
-		quizB := &quiz.Quiz{Title: "Other", Slug: "other", CreatedByPlayerID: seededAdminID}
-		if err := quizStore.CreateQuiz(ctx, quizB); err != nil {
-			t.Fatalf("CreateQuiz B err = %v, want nil", err)
-		}
-		groupOnB := firstRoundID(t, quizB, quizStore)
-
-		testGame := newTestGame(t, testQuiz)
-		if err := gameStore.CreateGame(ctx, testGame); err != nil {
-			t.Fatalf("CreateGame err = %v, want nil", err)
-		}
-		if err := gameStore.CreateParticipant(
-			ctx,
-			&Participant{GameID: testGame.ID, PlayerID: 1, QuizID: testQuiz.ID},
-		); err != nil {
-			t.Fatalf("CreateParticipant err = %v, want nil", err)
-		}
-
-		svc := NewService(gameStore, quizStore, slog.Default())
-		err := svc.MarkRoundSeen(ctx, testGame.ID, 1, groupOnB, RoundPhaseIntro)
-		if got, want := err, quiz.ErrRoundNotFound; !errors.Is(got, want) {
-			t.Errorf("err = %v, want %v", got, want)
 		}
 	})
 }
