+++ internal/store/quiz_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package store_test
 
 import (
